v4 - normalize project structure
