+++ main.go
@@ -3,47 +3,12 @@ package main
 import (
 	"fmt"
 	"os"
-	"strings"
 
-	"github.com/hashmap-kz/rconf/internal/cmd"
-	"github.com/hashmap-kz/rconf/internal/runner"
-	"github.com/spf13/cobra"
+	"github.com/hashmap-kz/rconf/cmd"
 )
 
-// main function using Cobra CLI
 func main() {
-	var cfg cmd.Config
-
-	rootCmd := &cobra.Command{
-		Use:   "rconf",
-		Short: "Execute local scripts on remote hosts via SSH",
-		Run: func(_ *cobra.Command, _ []string) {
-			runner.Run(&cfg)
-		},
-	}
-
-	rootCmd.Flags().StringVarP(&cfg.PrivateKeyPath, "pkey", "i", "", "Path to SSH private key (required when pkey-auth is used)")
-	rootCmd.Flags().StringVarP(&cfg.PrivateKeyPassphrase, "pkey-pass", "", "", "Passphrase to SSH private key (required when pkey is password-protected)")
-	rootCmd.Flags().StringSliceVarP(&cfg.Filenames, "filename", "f", nil, "List of script paths or directories (required)")
-	rootCmd.Flags().StringSliceVarP(&cfg.ConnStrings, "conn", "H", nil, strings.TrimSpace(`
-List of remote hosts (required)
-Format: username:password@host:port
-- password is optional
-- port is optional (default 22)
-`))
-	rootCmd.Flags().IntVarP(&cfg.WorkerLimit, "workers", "w", 2, "Max concurrent SSH connections")
-	rootCmd.Flags().StringVarP(&cfg.LogFile, "log", "l", "ssh_execution.log", "Log file path")
-	rootCmd.Flags().BoolVarP(&cfg.Recursive, "recursive", "R", true, "Process the directory used in -f, --filename recursively")
-
-	requiredFlags := []string{"filename", "conn"}
-	for _, flag := range requiredFlags {
-		if err := rootCmd.MarkFlagRequired(flag); err != nil {
-			fmt.Printf("Failed to mark '%s' flag as required: %v", flag, err)
-			os.Exit(1)
-		}
-	}
-
-	if err := rootCmd.Execute(); err != nil {
+	if err := cmd.Execute(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
